+++ Packages/com.unity.inputsystem/InputSystem/Editor/UITKAssetEditor/Views/InputActionsTreeViewItem.cs
@@ -3,7 +3,6 @@
 #if UNITY_EDITOR && UNITY_INPUT_SYSTEM_PROJECT_WIDE_ACTIONS
 using System.Threading.Tasks;
 using UnityEditor;
-using UnityEngine.InputSystem.Editor;
 using UnityEngine.UIElements;
 
 namespace UnityEngine.InputSystem.Editor
@@ -14,13 +13,9 @@ namespace UnityEngine.InputSystem.Editor
     internal class InputActionsTreeViewItem : VisualElement
     {
         public EventCallback<string> EditTextFinishedCallback;
-        public EventCallback<int> DeleteCallback;
-        public EventCallback<int> DuplicateCallback;
 
         private const string kRenameTextField = "rename-text-field";
         public event EventCallback<string> EditTextFinished;
-        public event EventCallback<int> OnDeleteItem;
-        public event EventCallback<int> OnDuplicateItem;
 
         private bool m_IsEditing;
 
@@ -100,16 +95,6 @@ namespace UnityEngine.InputSystem.Editor
         {
             await Task.Delay(120);
             renameTextfield.Q<TextField>().Focus();
-        }
-
-        public void DeleteItem()
-        {
-            OnDeleteItem?.Invoke(0);
-        }
-
-        public void DuplicateItem()
-        {
-            OnDuplicateItem?.Invoke(0);
         }
 
         private void OnEditTextFinished()